pyproject.toml
[tool.poetry]
name = 'catalog_bot'
version = '0.1.0'
description = ''
authors = []

[tool.black]
line-length = 100

[tool.isort]
atomic = true
profile = 'black'

[tool.mypy]
python_version = 3.11
disallow_untyped_defs = true
ignore_missing_imports = true
disable_error_code = 'arg-type'
no_namespace_packages = true

[tool.poetry.dependencies]
python = '3.11.4'
pre-commit = '^3.4.0'
punq = '0.7.0'
aiogram = '3.4.1'
django = '4.2.10'
djangorestframework = '3.14.0'
pillow = '10.2.0'
psycopg2-binary = '2.9.9'
aiogram-dialog = '^2.1.0'
gunicorn = "^21.2.0"
celery = "^5.3.6"
django-redis = "^5.4.0"
redis = "^5.0.2"

[tool.poetry.group.dev.dependencies]
flake8 = '^6.1.0'
flake8-bugbear = '^23.3.12'
flake8-builtins = '^2.1.0'
flake8-cognitive-complexity = '^0.1.0'
flake8-commas = '^2.1.0'
flake8-django = '^1.1.5'
flake8-quotes = '^3.3.2'
flake8-todo = '^0.7'
flake8-variables-names = '^0.0.5'
isort = '^5.12.0'
black='^23.7.0'
add-trailing-comma='^3.0.1'
mypy = '^1.8.0'
mypy-extensions = '^1.0.0'
typing-extensions = '^4.9.0'
